infrastruct/utils/requirements.txt
google-generativeai
tqdm
pathspec
